+++ src/Options/NatsTransportConfigurationBuilder.php
@@ -51,6 +51,7 @@ final class NatsTransportConfigurationBuilder
         TransportOption::STREAM_REPLICAS->value => 1,
         TransportOption::RETRY_HANDLER->value => RetryHandler::SYMFONY->value,
         TransportOption::SCHEDULED_MESSAGES->value => false,
+        TransportOption::ACK_SYNC->value => false,
         TransportOption::TLS_REQUIRED->value => false,
         TransportOption::TLS_HANDSHAKE_FIRST->value => false,
         TransportOption::TLS_CA_FILE->value => null,
@@ -112,6 +113,7 @@ final class NatsTransportConfigurationBuilder
             options: $configuration,
             natsRetryHandlerEnabled: $configuration[TransportOption::RETRY_HANDLER->value] === RetryHandler::NATS->value,
             scheduledMessagesEnabled: $this->toBool($configuration[TransportOption::SCHEDULED_MESSAGES->value]),
+            ackSyncEnabled: $this->toBool($configuration[TransportOption::ACK_SYNC->value]),
         );
     }