+++ src/Options/TransportOption.php
@@ -14,6 +14,7 @@ namespace IDCT\NatsMessenger\Options;
  * **Stream Limits:** STREAM_MAX_AGE, STREAM_MAX_BYTES, STREAM_MAX_MESSAGES,
  *                    STREAM_MAX_MESSAGES_PER_SUBJECT, STREAM_STORAGE, STREAM_REPLICAS
  * **Retry Strategy:** RETRY_HANDLER
+ * **Acknowledgement:** ACK_SYNC
  * **Scheduling:** SCHEDULED_MESSAGES
  * **TLS:** TLS_REQUIRED, TLS_HANDSHAKE_FIRST, TLS_CA_FILE, TLS_CERT_FILE, TLS_KEY_FILE,
  *          TLS_KEY_PASSPHRASE, TLS_PEER_NAME, TLS_VERIFY_PEER
@@ -33,6 +34,7 @@ enum TransportOption: string
     case STREAM_REPLICAS = 'stream_replicas';
     case RETRY_HANDLER = 'retry_handler';
     case SCHEDULED_MESSAGES = 'scheduled_messages';
+    case ACK_SYNC = 'ack_sync';
 
     case TLS_REQUIRED = 'tls_required';
     case TLS_HANDSHAKE_FIRST = 'tls_handshake_first';